+++ test/Test/Fixtures/oracle/UnboxedTuples/unboxed-tuple-type-constructor.hs
@@ -1,0 +1,7 @@
+{- ORACLE_TEST pass -}
+{-# LANGUAGE UnboxedTuples #-}
+module UnboxedTupleTypeConstructor where
+
+type P = (# , #) Int Bool
+
+type T = (# , , #)